% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_sig_subpath.R
\name{all_sig_subpath}
\alias{all_sig_subpath}
\title{Table of Significant Subpathways}
\usage{
all_sig_subpath(path_results)
}
\arguments{
\item{path_results}{Results data frame generated by \code{\link{subpathway_analysis}}}
}
\value{
A table of all significant subpathways. Including the significant model type
and model type p-value.
}
\description{
Create a table of all significant subpathways
}
