% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_heatmap_Data.R
\name{create_heatmap_Data}
\alias{create_heatmap_Data}
\title{Create metadata and matricies for metabolite heatmaps}
\usage{
create_heatmap_Data(.analysis_data, heatmap_variables, ...)
}
\arguments{
\item{.analysis_data}{The analysis data containing all of the metabolite and
metadata information.}

\item{heatmap_variables}{A vector of variable names that are NOT metabolites.}

\item{...}{Additional arguments that can be passed into the arrange function.
This parameter will order the columns of the heatmap data.}
}
\value{
A list of matricies including the heatmap variable (meta data for heatmap)
and the values for the heatmap.
}
\description{
This function creates the required matrices for the metabolite heatmaps.
}
