% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMetPipe.R
\name{createMetPipe}
\alias{createMetPipe}
\title{Manually Create MetPipe Object}
\usage{
createMetPipe(
  raw_peak = data.frame(),
  standardized_peak = data.frame(),
  meta = data.frame(),
  chemical_annotation = data.frame()
)
}
\arguments{
\item{raw_peak}{Raw peak data}

\item{standardized_peak}{standardized peak data}

\item{meta}{Sample meta data}

\item{chemical_annotation}{Chemical annotation data}
}
\value{
MetPipe

The MetPipe data object contains all of the information needed to complete downstream
analysis.
}
\description{
Create a MetPipe object from metabolomics matricies.
}
