% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demoDataSmall.R
\name{demoDataSmall}
\alias{demoDataSmall}
\title{Subset of Demo data for the MetabolomicsPipeline ,}
\format{
Rd
}
\description{
Demo data consisting of 86 samples (42 males, 44 females), three treatment groups, and the samples were taken
}
\keyword{data}
