% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demoSampleMeta.R
\docType{data}
\name{demoSampleMeta}
\alias{demoSampleMeta}
\title{Demo sample metadata for the MetabolomicsPipeline,}
\format{
data.frame
}
\usage{
data(demoSampleMeta)
}
\value{
A dataframe with metadata for with 86 samples.
}
\description{
Demo data consisting of 86 samples (42 males, 44 females), three treatment
groups, and the samples were taken
}
\keyword{data}
