% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadMetabolon.R
\name{loadMetabolon}
\alias{loadMetabolon}
\title{Load Metabolomic Data as SummarizedExperiment}
\usage{
loadMetabolon(
  path,
  raw_sheet = "Peak Area Data",
  chemical_sheet = "Chemical Annotation",
  sample_meta = "Sample Meta Data",
  normalized_peak = "Log Transformed Data",
  sample_names = "PARENT_SAMPLE_NAME",
  chemicalID = "CHEM_ID"
)
}
\arguments{
\item{path}{Path to excel file with peak data, chemical
annotations, sample meta data, and (optionally) the normalized peak counts}

\item{raw_sheet}{Sheet name for the raw peak data.}

\item{chemical_sheet}{Sheet name for chemical annotation.}

\item{sample_meta}{Sheet name for sample meta data.}

\item{normalized_peak}{Sheet name for the normalized peak data.
If you are not adding the normalized data from the excel file then set
normalized_peak=NA.}

\item{sample_names}{Column name in the meta data containing the sample names.
This must correspond to the row names of the raw peak data in the excel
file.}

\item{chemicalID}{Column name in the meta data containing the sample names.
This must correspond to the column names of the raw peak data.}
}
\value{
A SummarizedExperiment containing metabolomics expirement data.
}
\description{
Automatically load metabolomic data from excel file
}
\details{
The metabolomics experiment data are stored in a SummarizedExperiment.
}
\seealso{
\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment::SummarizedExperiment}
}
