% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_transformation.R
\name{log_transformation}
\alias{log_transformation}
\title{Log transformation of metabolite data}
\usage{
log_transformation(peak_data)
}
\arguments{
\item{peak_data}{A matrix of peak data with metabolites in the columns}
}
\value{
log transformed peak data
}
\description{
This function log transforms each metabolite in the Metabolon data.
}
\examples{
data("demoDataSmall", package = "MetabolomicsPipeline")
peak <- assay(demoDataSmall, "peak")

# Median standardization
peak_med <- median_standardization(peak_data = peak)

# Min value imputation
peakImpute <- min_val_impute(peak_data = peak_med)

# log transformation
peak_log <- log_transformation(peak_data = peakImpute)


}
