% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median_standardization.R
\name{median_standardization}
\alias{median_standardization}
\title{Median standardization for metabolite data}
\usage{
median_standardization(peak_data)
}
\arguments{
\item{peak_data}{Peak data with metabolites in the columns. The data also
must include the "Parent_SAMPLE_NAME".}
}
\value{
Median standardized peak data.
}
\description{
This function standardizes the metabolites by the median of the metabolite.
}
