% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medianStandardization.R
\name{medianStandardization}
\alias{medianStandardization}
\title{Median standardization for metabolite data}
\usage{
medianStandardization(met_se, assay = "peak")
}
\arguments{
\item{met_se}{A \code{SummarizedExperiment} object containing metabolite data.}

\item{assay}{A character string specifying the assay name
within \code{met_se} to be median standardized.
Default is \code{"peak"}.}
}
\value{
A data frame of median standardized metabolite data.
The \code{SummarizedExperiment} object is also updated
internally with a new assay \code{"median_std"}.
}
\description{
This function performs median standardization of metabolite
data stored within a \code{SummarizedExperiment} object.
For each metabolite, the values are divided by the
median value of that metabolite across samples.
The standardized data are returned as a data frame
and also added to the original \code{SummarizedExperiment}
object as a new assay named \code{"median_std"}.
}
\examples{
library(SummarizedExperiment)
data("demoDataSmall", package = "MetabolomicsPipeline")

# Median standardization
peak_med <- medianStandardization(met_se = demoDataSmall, assay = "peak")

# Access the median standardized data within the SummarizedExperiment
assay(peak_med, "median_std")[1:5, 1:5]

}
