% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met_within_sub.R
\name{met_within_sub}
\alias{met_within_sub}
\title{Metabolites within Subpathway Table}
\usage{
met_within_sub(
  path_results,
  subpathway,
  mod = c("interaction", "parallel", "single")
)
}
\arguments{
\item{path_results}{Results data frame generated by \code{\link{subpathway_analysis}}}

\item{subpathway}{Character string of the subpathway of interest. This is case sensitive
and must be listed in the path_results.}

\item{mod}{model of interest. This can be a single model or a vector of model types
that can take on the values interaction,parallel, or single.}
}
\value{
a table with the results from the model types specified and for each
metabolite within the superpathway specified.
}
\description{
Return the test results for each metabolite within a subpathway.
}
