% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolite_pairwise.R
\name{metabolite_pairwise}
\alias{metabolite_pairwise}
\title{Metabolite Pairwise Comparisons.}
\usage{
metabolite_pairwise(MetPipe, form, strat_var = NULL)
}
\arguments{
\item{form:}{This is a character string the resembles the right hand side of a
simple linear regression model in R. For example form = "Group1 + Group2".}

\item{Metpipe:}{The analysis data we will use for the pairwise comparisons. The
data must be organized in a MetPipe object.}

\item{strat_var:}{A variable in the analysis data to stratify the model by. If this
is specified, a list of results will be returned.}
}
\value{
The overall F-test p-value, and the estimate and pvalue for each pairwise comparison.
}
\description{
Computes the pairwise comparison estimates and p-values for each metabolite.
}
\details{
This function will analyze each metabolite individually. For each metabolite, the metabolite_pairwise function
will first test whether the model explained a significant proportion
of the variance in the metabolite using an F-test. Since we will be looking at
multiple comparisons for the metabolite, it is good practice to first look at the
overall p-value from the F-test before looking at the pairwise comparisons.
The metabolite_pairwise function then looks at all pairwise comparisons utilizing
the \href{https://cran.r-project.org/web/packages/emmeans/index.html}{emmeans} package.
The metabolite_pairwise function returns a data frame with the metabolite overall
p-value, log fold change for each group, and the p-value
for each comparison.
}
