% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolite_pca.R
\name{metabolite_pca}
\alias{metabolite_pca}
\title{Metabolite PCA}
\usage{
metabolite_pca(MetPipe, meta_var)
}
\arguments{
\item{meta_var}{A metadata variable to color code the PCA plot by.}

\item{analysis_data}{\itemize{
\item analysis data with metabolites in the columns
}}

\item{A}{vector of metabolites to include in the PCA. This is primarily to
to remove columns that do not include metabolite peak data.}
}
\value{
A PCA plot of the first two principal components, colored by the metadata
variable.
}
\description{
Computes and plots the first two components of the PCA from metabolite data.
}
