% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolitePairwise.R
\name{metabolitePairwise}
\alias{metabolitePairwise}
\title{Metabolite Pairwise Comparisons.}
\usage{
metabolitePairwise(
  data,
  form,
  Assay = "normalized",
  strat_var = NULL,
  mets = NULL
)
}
\arguments{
\item{data}{SummarizedExperiment with metabolomics experiment data.}

\item{form}{This is a character string the resembles the right hand side of
a simple linear regression model in R. For example form = "Group1 + Group2".}

\item{Assay}{Name of the assay to be used for the pairwise analysis
(default='normalized')}

\item{strat_var}{A variable in the analysis data to stratify the model by.
If this is specified, a list of results will be returned.}

\item{mets}{Chemical ID for the metabolites of interest. If NULL then the
pairwise analysis is completed for all metabololites.}
}
\value{
The overall F-test p-value, and the estimate and pvalue for each
pairwise comparison.
}
\description{
Computes the pairwise comparison estimates and p-values for each metabolite.
}
\details{
This function will analyze each metabolite individually. For each
metabolite, the metabolite_pairwise function
will first test whether the model explained a significant proportion
of the variance in the metabolite using an F-test. Since we will be looking
at multiple comparisons for the metabolite, it is good practice to first look
at the overall p-value from the F-test before looking at the pairwise
comparisons.
The metabolite_pairwise function then looks at all pairwise comparisons
utilizing the
\href{https://cran.r-project.org/web/packages/emmeans/index.html}{emmeans}
package. The metabolite_pairwise function returns a data frame with the
metabolite overall p-value, log fold change for each group, and the p-value
for each comparison.
}
\examples{
# Load data
data("demoDataSmall", package = "MetabolomicsPipeline")
dat <- demoDataSmall

# Run pairwise analysis
stratPairwise <- metabolite_pairwise(dat,
    form = "GROUP_NAME*TIME1",
    strat_var = "Gender"
)

#############################################################################
## Create Estimate Heatmap #####################################################
################################################################################

metEstHeatmap(strat_pairwise$Female, dat,
               interactive = FALSE,
               CHEM_ID = "CHEM_ID", SUB_PATHWAY = "SUB_PATHWAY",
               CHEMICAL_NAME = "CHEMICAL_NAME",
               main = "Log fold change heatmap", show_rownames = FALSE
)


################################################################################
## Create P-value Heatmap ######################################################
################################################################################
# Female
metPHeatmap(strat_pairwise$Female, dat,
             interactive = FALSE, show_rownames = FALSE,
             main = "Pvalue Heatmap"
)

}
