% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolitePca.R
\name{metabolitePca}
\alias{metabolitePca}
\title{Metabolite PCA}
\usage{
metabolitePca(data, Assay = "normalized", meta_var)
}
\arguments{
\item{data}{SummarizedExperiment with metabolomics experiment data.}

\item{Assay}{Name of the assay to be used for the pairwise analysis
(default='normalized')}

\item{meta_var}{A metadata variable to color code the PCA plot by.}
}
\value{
A PCA plot of the first two principal components, colored by the
metadata variable.
}
\description{
Computes and plots the first two components of the PCA from the metabolite
data.
}
\examples{

# load data
data("demoDat", package = "MetabolomicsPipeline")
dat <- demoDat

# Define PCA label from metadata
meta_var <- "Gender"

# Run PCA
pca <- metabolitePca(dat,
    meta_var = meta_var
)


# Show PCA
pca

}
