% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetabolomicsPipeline-package.R
\docType{package}
\name{MetabolomicsPipeline-package}
\alias{MetabolomicsPipeline}
\alias{MetabolomicsPipeline-package}
\title{MetabolomicsPipeline: Metabolomics Pipeline Tools}
\description{
This package provides analysis tools for analyzing metabolomics data. We provide functionality for hypothesis testing at the subpathway level, pairwise comparisons of metabolites, and tools for exploratory analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/datalifecycle-ua/MetabolomicsPipeline}
  \item Report bugs at \url{https://github.com/datalifecycle-ua/MetabolomicsPipeline/issues}
}

}
\author{
\strong{Maintainer}: Joel Parker \email{joelparker@arizona.edu} (\href{https://orcid.org/0000-0002-3411-3818}{ORCID})

Authors:
\itemize{
  \item Bonnie LaFleur \email{blafleur@arizona.edu}
}

}
\keyword{internal}
