% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetLabClass.R
\docType{methods}
\name{MetPipe}
\alias{MetPipe}
\alias{MetLab}
\title{Define metlab class}
\description{
Define metlab class
}
\section{Slots}{

\describe{
\item{\code{raw_peak}}{Raw Peak Data}

\item{\code{standardized_peak}}{Standardized Peak data}

\item{\code{meta}}{Sample Metadata}

\item{\code{chemical_annotation}}{Chemical Annotation file}

\item{\code{analysis}}{Analysis data file}
}}

