% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_val_impute.R
\name{min_val_impute}
\alias{min_val_impute}
\title{Minimum Value Imputation}
\usage{
min_val_impute(peak_data)
}
\arguments{
\item{peak_data}{Peak data matrix with metabolites in the columns.}
}
\value{
Metabolite imputed peak data.
}
\description{
Imputes the minimum value for each metabolite
}
