% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minValImpute.R
\name{minValImpute}
\alias{minValImpute}
\title{Minimum Value Imputation for Metabolite Data}
\usage{
minValImpute(met_se, assay = "median_std")
}
\arguments{
\item{met_se}{A \code{SummarizedExperiment} object containing metabolite data.}

\item{assay}{A character string specifying
the assay name within \code{met_se} to perform minimum value imputation on.
Default is \code{"median_std"}.}
}
\value{
The input \code{SummarizedExperiment} object with a
new assay \code{"min_impute"} containing the minimum value-imputed data.
}
\description{
This function imputes missing values in metabolite
data stored within a \code{SummarizedExperiment} object.
For each metabolite, missing values are replaced
with the minimum observed value for that metabolite.
The imputed data are added to the \code{SummarizedExperiment}
object as a new assay named \code{"min_impute"}.
}
\examples{
library(SummarizedExperiment)
data("demoDataSmall", package = "MetabolomicsPipeline")

# Median standardization
demoDataSmall <- median_standardization(met_se = demoDataSmall,
 assay = "peak")

# Minimum value imputation
demoDataSmall <- min_val_impute(met_se = demoDataSmall, assay = "median_std")

# Access the imputed data
assay(demoDataSmall, "min_impute")[1:5, 1:5]

}
