% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subpath_within_superpath.R
\name{subpath_within_superpath}
\alias{subpath_within_superpath}
\title{Proportion of the Significant Subpathways Within Superpathways}
\usage{
subpath_within_superpath(subpath_results)
}
\arguments{
\item{subpath_results}{Results data frame generated by
\code{\link{subpathway_analysis}}}
}
\value{
A table with the proportion (and percent) of significant subpathways
within superpathways.
}
\description{
Create a table that gives the percentage of significant subpathways within
each superpathway.
}
\examples{

# Load data
data("demoDataSmall", package = "MetabolomicsPipeline")
dat <- demoDataSmall

# Runsubpathay analysis
sub_analysis = subpathway_analysis(dat,
                                  treat_var = "GROUP_NAME",
                                  block_var = "TIME1",
                                  strat_var = NULL,
                                  Assay = "normalized")

#############################################################################
### Results Plots ###########################################################
##############################################################################

# significant subpathways by model type
subpath_by_model(sub_analysis)

# Percentage of signficant subpathways within superpathways
subpath_within_superpath(sub_analysis)

met_within_sub(sub_analysis, subpathway = "Aminosugar Metabolism")




}
