% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subpathway_analysis.R
\name{subpathway_analysis}
\alias{subpathway_analysis}
\title{Subpathway Analysis}
\usage{
subpathway_analysis(data, chem_data, treat_var, block_var)
}
\arguments{
\item{data}{The analysis data containing sample meta data and metabolite peak data}

\item{chem_data}{Chemical Annotation data provided by Metabolon.}

\item{treat_var}{This is the name of the variable in the analysis data that is the main variable of interest.}

\item{block_var}{Is the name of the blocking variable in the dataset.
If the the experimental design does not include a blocking variable, then the
value of block_var=NULL.}
}
\value{
A data frame with "CHEM_ID","sub_pathway","chem_name","interaction_pval","interaction_fisher","parallel_pval","parallel_fisher","single_pval","single_fisher",and "model" for each metabolite.
}
\description{
Subpathway analysis for metabolite data.
}
\details{
For each metabolite, we test three models using using ANOVA.
\itemize{

\item{Interaction:}{ \eqn{log Peak = Treatment + block + Treatment*block}}

\item{Parallel: }{ \eqn{log Peak = Treatment + block}}

\item{Single: }{ \eqn{log Peak =  Treatment}}

}

For the interaction model, we are focusing only on the interaction term
"Treatment*block" for this to test if there is a  significant interaction between our
treatment and the block variable. The parallel model is testing if the block
variable is explaining a significant amount of the metabolite variance,
and the treatment model is testing if the treatment explains a
significant proportion of the variance for each metabolite.
Then, each model will be tested at the subpathway level with the Combined Fisher Probability.
\deqn{\tilde{X} = -2\sum_{i=1}^k ln(p_i)}
}
\seealso{
\href{https://www.sciencedirect.com/science/article/pii/S0167947303002950}{Loughin, Thomas M. "A systematic comparison of methods for combining p-values from independent tests." Computational statistics & data analysis 47.3 (2004): 467-485.}
}
