% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subpathway_boxplots.R
\name{subpathway_boxplots}
\alias{subpathway_boxplots}
\title{Subpathway Boxplots}
\usage{
subpathway_boxplots(analysis_data, chem_data, subpathway, X, groupBy, ...)
}
\arguments{
\item{analysis_data}{The analysis data containing sample meta data and metabolite peak data.}

\item{chem_data}{Chemical Annotation data provided by Metabolon.}

\item{subpathway}{Character value of the subpathway of interest. This is case
sensitive and must be in the chem_data.}

\item{X}{This the the name of the variable in the meta data that is used for the
X axis of the box plots.}

\item{groupBy}{This is a grouping variable. As a recommendation the treatment
groups should be used in the groupBy argument as this will provide a different color
for each of the treatments making it easier to identify.}
}
\value{
Boxplots stratified by metabolites.
}
\description{
Creates boxplots for each metabolite within a specified subpathway.
}
