% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subpathWithinSuperpath.R
\name{subpathWithinSuperpath}
\alias{subpathWithinSuperpath}
\title{Proportion of the Significant Subpathways Within Superpathways}
\usage{
subpathWithinSuperpath(subpath_results)
}
\arguments{
\item{subpath_results}{Results data frame generated by
\code{\link{subpathway_analysis}}}
}
\value{
A table with the proportion (and percent) of significant subpathways
within superpathways.
}
\description{
Create a table that gives the percentage of significant subpathways within
each superpathway.
}
\examples{

# Load data
data("demoDataSmall", package = "MetabolomicsPipeline")
dat <- demoDataSmall

# Runsubpathay analysis
sub_analysis <- subpathwayAnalysis(dat,
    treat_var = "GROUP_NAME",
    block_var = "TIME1",
    strat_var = NULL,
    Assay = "normalized"
)

#############################################################################
### Results Plots ###########################################################
#############################################################################

# significant subpathways by model type
subpathByModel(sub_analysis)

# Percentage of signficant subpathways within superpathways
subpathWithinSuperpath(sub_analysis)

met_within_sub(sub_analysis, subpathway = "Aminosugar Metabolism")

}
