% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_sig_subpath.R
\name{all_sig_subpath}
\alias{all_sig_subpath}
\title{Table of Significant Subpathways}
\usage{
all_sig_subpath(path_results)
}
\arguments{
\item{path_results}{Results data frame generated by
\code{\link{subpathway_analysis}}}
}
\value{
A table of all significant subpathways. Including the significant
model type
and model type p-value.
}
\description{
Create a table of all significant subpathways
}
\examples{
# Load data
data("demoDataSmall", package = "MetabolomicsPipeline")
dat <- demoDataSmall

# Runsubpathay analysis
sub_analysis <- subpathway_analysis(dat,
                                  treat_var = "GROUP_NAME",
                                  block_var = "TIME1",
                                  strat_var = NULL,
                                  Assay = "normalized")

##############################################################################
### Results Plots ############################################################
##############################################################################

# significant subpathways by model type
subpath_by_model(sub_analysis)

# Percentage of signficant subpathways within superpathways
subpath_within_superpath(sub_analysis)

met_within_sub(sub_analysis, subpathway = "Aminosugar Metabolism")

# All signifiicant subpathways
all_sig_subpath(sub_analysis)

}
