% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_met_se.R
\name{create_met_se}
\alias{create_met_se}
\title{Create a SummarizedExperiment from   Metabolomics Data}
\usage{
create_met_se(
  chemical_annotation,
  sample_metadata,
  peak_data,
  sample_names = "PARENT_SAMPLE_NAME",
  chemical_id = "CHEM_ID"
)
}
\arguments{
\item{chemical_annotation}{A \code{data.frame} where each row represents a chemical.}

\item{sample_metadata}{A \code{data.frame} where each row represents a biological sample.}

\item{peak_data}{A \code{data.frame} with samples as rows and chemicals as columns.
Sample names must be the first column.}

\item{sample_names}{Column name in the meta data containing the sample names.
This must correspond to the row names of the raw peak data in the excel
file.}

\item{chemical_id}{Column name in the meta data containing the sample names.
This must correspond to the column names of the raw peak data.}
}
\value{
A \code{SummarizedExperiment} object containing:
\itemize{
\item assays: the peak data
\item rowData: the chemical annotation
\item colData: the sample metadata
}
}
\description{
This function constructs a \code{SummarizedExperiment}
object from   metabolomics data.
It uses peak intensity data,
sample metadata as \code{colData},
and chemical annotations as \code{rowData}.
All inputs must be aligned via identifiers (sample_names and chemicalID)
}
\examples{
#Get demo sample metadata
data("demoSampleMeta", package = "MetabolomicsPipeline")

#Get demo chemical annotation file
data("demoChemAnno", package = "MetabolomicsPipeline")

# Get demo peak data
data("demoPeak", package = "MetabolomicsPipeline")

dat <- create_met_se(chemical_annotation = demoChemAnno,
                    sample_metadata = demoSampleMeta,
                    peak_data = demoPeak,
                    chemical_id = "CHEM_ID",
                    sample_names = "PARENT_SAMPLE_NAME")

}
