% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demoDat.R
\docType{data}
\name{demoDat}
\alias{demoDat}
\title{Demo data for the MetabolomicsPipeline,}
\format{
SummarizedExperiment object
}
\usage{
data(demoDat)
}
\value{
A SummarizedExperiment object with 86 samples
}
\description{
Demo data consisting of 86 samples (42 males, 44 females), three treatment
groups, and the samples were taken
}
\keyword{data}
