% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demoDataSmall.R
\docType{data}
\name{demoDataSmall}
\alias{demoDataSmall}
\title{Subset of Demo data for the MetabolomicsPipeline ,}
\format{
Rd
}
\usage{
data(demoDataSmall)
}
\value{
A subset of the metabolites in the DemoData.
}
\description{
Demo data consisting of 86 samples (42 males, 44 females), three treatment
groups, and the samples were taken at three different time points.
We focus on a subset of 10 Subpathways.
}
\keyword{data}
