% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_transformation.R
\name{log_transformation}
\alias{log_transformation}
\title{Log transformation of metabolite data}
\usage{
log_transformation(peak_data)
}
\arguments{
\item{peak_data}{A matrix of peak data with metabolites in the columns}
}
\value{
log transformed peak data
}
\description{
This function log transforms each metabolite in the Metabolon data.
}
