% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met_p_heatmap.R
\name{met_p_heatmap}
\alias{met_p_heatmap}
\title{Metabolite Pairwise P-Value Interactive Heatmap.}
\usage{
met_p_heatmap(results_data, chem_data)
}
\arguments{
\item{results_data}{Results data frame of the pairwise comparisons produced by \code{\link{metabolite_pairwise}}.}

\item{chem_data}{The chemical annotation data provided by Metabolome}
}
\value{
An interactive heatmap of pairwise pvalues.

#' @details
For the metabolites which had a significant overall p-value (which tested if the
treatment group means were equal under the null hypothesis), we will produce a
heatmap of the pairwise p-values.
}
\description{
Produce an interactive heatmap of the pvalues produced in \code{\link{metabolite_pairwise}}.
}
\details{
For the metabolites which had a significant overall p-value (which tested if the
treatment group means were equal under the null hypothesis), we will produce a
heatmap of the p-values.
}
