% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met_within_sub.R
\name{met_within_sub}
\alias{met_within_sub}
\title{Metabolites within Subpathway Table}
\usage{
met_within_sub(
  subpath_results,
  subpathway,
  mod = c("interaction", "parallel", "single")
)
}
\arguments{
\item{subpath_results}{Results data frame generated by \code{\link{subpathway_analysis}}}

\item{subpathway}{Character string of the subpathway of interest. This is case sensitive
and must be listed in the subpath_results.}

\item{mod}{Model of interest. This can be a single model or a vector of model types
that can take on the values "interaction", "parallel", or "single".}
}
\value{
A table with the results from the model types specified and for each
metabolite within the superpathway specified.
}
\description{
Return the model results for each metabolite within a subpathway.
}
\examples{
data("demoDataSmall", package = "MetabolomicsPipeline")
dat <- demoDataSmall

# Runsubpathay analysis
sub_analysis = subpathway_analysis(dat,
                                  treat_var = "GROUP_NAME",
                                  block_var = "TIME1",
                                  strat_var = NULL,
                                  Assay = "normalized")

################################################################################
### Results Plots ##############################################################
################################################################################

# significant subpathways by model type
subpath_by_model(sub_analysis)

# Percentage of signficant subpathways within superpathways
subpath_within_superpath(sub_analysis)

met_within_sub(sub_analysis, subpathway = "Aminosugar Metabolism")


}
