% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolite_heatmap.R
\name{metabolite_heatmap}
\alias{metabolite_heatmap}
\title{Create metabolite heatmap}
\usage{
metabolite_heatmap(.analysis_data, heatmap_variables, metabolites, ...)
}
\arguments{
\item{.analysis_data}{The analysis data containing all of the metabolite and
metadata information.}

\item{heatmap_variables}{A vector of variable names that are NOT metabolites to include
in the heatmap.}

\item{metabolites}{a vector of metabolite names to include in the heatmap.}

\item{...}{Additional arguments that can be passed into the arrange function.
This parameter will order the columns of the heatmap.}
}
\value{
Heatmap plot with the columns arranged as specified in ...
}
\description{
This function creates metabolite heatmaps
}
