% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolite_heatmap.R
\name{metabolite_heatmap}
\alias{metabolite_heatmap}
\title{Create metabolite heatmap}
\usage{
metabolite_heatmap(
  data,
  top_mets = 50,
  group_vars,
  strat_var = NULL,
  caption = NULL,
  Assay = "normalized",
  ...
)
}
\arguments{
\item{data}{A SummarizedExperiment containing the Metabolon experiment data.}

\item{top_mets}{Number of metabolites to include in the heatmap. Metabolites
are chosen based on the highest variability.}

\item{group_vars}{Vector of variables to annotate heatmap with. Columns will
be grouped by these variables.}

\item{strat_var}{Variable to stratify the heatmap by.}

\item{caption}{A title for the heatmap. If strat_var is used, the title will
automatically include the stratum with the tile.}

\item{Assay}{Which assay data to use for the heatmap (default="normalized").}

\item{...}{Additional arguments can be passed into the arrange function.
This parameter will order the columns of the heatmap.}
}
\value{
A gtable class with all of the information to build the heatmap. To view
the heatmap use ggplotify::as.ggplot().
}
\description{
Create heatmaps which are arranged by the experimental conditions.
}
