% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolite_pairwise.R
\name{metabolite_pairwise}
\alias{metabolite_pairwise}
\title{Metabolite Pairwise Comparisons.}
\usage{
metabolite_pairwise(form, data, metabolites)
}
\arguments{
\item{form:}{This is a character string the resembles the right hand side of a
simple linear regression model in R. For example form = "Group1 + Group2".}

\item{data:}{The analysis data we will use for the pairwise comparisons.}

\item{Metabolites:}{A list of metabolites that we want to be analyzed.}
}
\value{
The overall F-test pvalue, and the estimate and pvalue for each pairwise comparison.
}
\description{
Computes the pairwise comparison estimates and pvalues for each metabolite.
}
\details{
This function will analyze each metabolite individually. For each metabolite, the the metabolite_pairwise function
will first test if any of the experimental groups explained a significant proportion
of the variance in the metabolite using an F-test. Since we will be looking at
multiple comparisons for the metabolite it is good practice to first look at the
over-all p-value from the F-test before looking at the pairwise comparisons.
The metabolite_pairwise function then looks at all pairwise comparisons utilizing
the \href{https://cran.r-project.org/web/packages/emmeans/index.html}{emmeans} package.
The metabolite_pairwise function returns a data frame with the metabolite, overall
p-value, estimated difference in means for all experimental group combinations, and the p-value
which test if the difference the groups is significantly different.
}
