% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetabolomicsPipeline-package.R
\docType{package}
\name{MetabolomicsPipeline-package}
\alias{MetabolomicsPipeline}
\alias{MetabolomicsPipeline-package}
\title{MetabolomicsPipeline: Metabolomics Pipeline Tools}
\description{
This package provides analysis tools for analyzing metabolomics data from Metabolon. The tools in this package compliment the analysis from Metabolon. We provide functionality for hypothesis testing at the subpathway level, pairwise comparisons of metabolites, and tools for exploratory analysis.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/JoelParkerUofA/MetabolomicsPipeline/issues}
}

}
\author{
\strong{Maintainer}: Bonnie Lafleur \email{blafleur@arizona.edu}

Authors:
\itemize{
  \item Joel Parker \email{joelparker@arizona.edu}
}

}
\keyword{internal}
