% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetabolomicsPipeline-package.R
\docType{package}
\name{MetabolomicsPipeline-package}
\alias{MetabolomicsPipeline}
\alias{MetabolomicsPipeline-package}
\title{MetabolomicsPipeline: Metabolomics Pipeline Tools}
\description{
This package was developed to provide analysis tools for analyzing metabolomics data produced by Metabolon.
}
\author{
\strong{Maintainer}: Bonnie Lafleur \email{blafleur@arizona.edu}

Authors:
\itemize{
  \item Joel Parker \email{joelparker@arizona.edu}
}

}
\keyword{internal}
