% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise.R
\name{pairwise}
\alias{pairwise}
\title{Pairwise function}
\usage{
pairwise(out, form, data)
}
\arguments{
\item{out}{Outcome used as reponse}

\item{form}{form of the model}

\item{data}{data used for modeling}
}
\description{
This is the main function for metabolite_pairwise
}
