% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subpath_by_model.R
\name{subpath_by_model}
\alias{subpath_by_model}
\title{Subpathway model type table}
\usage{
subpath_by_model(subpath_results)
}
\arguments{
\item{subpath_results}{Results data frame generated by \code{\link{subpathway_analysis}}}
}
\value{
A table of the number of significant subpathways by model type.
}
\description{
Create a table with the number of significant subpathways for each model type.
}
\details{
Each subpathway will only have one model type. We first test the interaction,
then the parellel, and the Single model is tested last. If a subpathway has
a significant interaction model type then the table will count it as a iteraction and
not as a parellel or single.
}
