% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subpath_within_superpath.R
\name{subpath_within_superpath}
\alias{subpath_within_superpath}
\title{Proportion of Significant Subpathways Within Superpathways}
\usage{
subpath_within_superpath(subpath_results)
}
\arguments{
\item{subpath_results}{Results data frame generated by \code{\link{subpathway_analysis}}}
}
\value{
A table with the proportion (and percent) of significant subpathways within superpathway.
}
\description{
Create a table that give the percentage of significant subpathways within each
superpathway.
}
\details{
The numbers reported are the number of subpathways that were significant for any
model.
}
