% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subpathwayBoxplots.R
\name{subpathway_boxplots}
\alias{subpathway_boxplots}
\title{Subpathway Boxplots}
\usage{
subpathway_boxplots(
  data,
  subpathway,
  block_var,
  treat_var,
  Assay = "normalized",
  CHEMICAL_NAME = "CHEMICAL_NAME",
  SUB_PATHWAY = "SUB_PATHWAY",
  ...
)
}
\arguments{
\item{data}{SummarizedExperiment with metabolomics experiment data.}

\item{subpathway}{Character value of the subpathway of interest. This is case
sensitive and must be in the chemical annotation file.}

\item{block_var}{This the the name of the variable in the meta data that is
used for the X axis of the box plots. We recommend using the "block_var"
from the subpathway analysis.}

\item{treat_var}{This is a grouping variable. As a recommendation the
treatment groups should be used in the treat_var argument as this will
provide a different color for each of the treatments making it easier to
identify.}

\item{Assay}{Name of the assay to be used for the pairwise analysis
(default='normalized')}

\item{CHEMICAL_NAME}{Column name in the chemical annotation worksheet which
contains the chemical name.}

\item{SUB_PATHWAY}{Column name in chemical annotation file which contains
the SUB_PATHWAY information}

\item{...}{Additional arguments to filter the analysis data by.}
}
\value{
Boxplots stratified by metabolites.
}
\description{
Creates boxplots for each metabolite within a specified subpathway.
}
\details{
.
}
\examples{
# load data
data("demoDat", package = "MetabolomicsPipeline")
dat <- demoDat

#############################################################################
### BoxPlots ###############################################################
############################################################################

subpathway_boxplots(dat,
    subpathway = "Lactoyl Amino Acid", block_var = TIME1,
    treat_var = GROUP_NAME, Assay = "normalized",
    CHEMICAL_NAME = "CHEMICAL_NAME",
    SUB_PATHWAY = "SUB_PATHWAY", Gender == "Female"
)


############################################################################
## Line plots ##############################################################
############################################################################

# Set up data
dat$TIME1 <- as.numeric(factor(dat$TIME1,
    levels = c("PreSymp", "Onset", "End")
))
# Create line plots
subpathway_lineplots(dat,
                    subpathway = "Lactoyl Amino Acid",
                    block_var = TIME1, treat_var = GROUP_NAME,
                    Assay = "normalized",
                    CHEMICAL_NAME = "CHEMICAL_NAME",
                    SUB_PATHWAY="SUB_PATHWAY",Gender == "Female")



}
