% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subpathway_boxplots.R
\name{subpathway_boxplots}
\alias{subpathway_boxplots}
\title{Subpathway Boxplots}
\usage{
subpathway_boxplots(
  data,
  subpathway,
  block_var,
  treat_var,
  Assay = "normalized",
  ...
)
}
\arguments{
\item{data}{SummarizedExperiment with Metabolon experiment data.}

\item{subpathway}{Character value of the subpathway of interest. This is case
sensitive and must be in the chemical annotation file.}

\item{block_var}{This the the name of the variable in the meta data that is used for the
X axis of the box plots. We recommend using the "block_var" from the subpathway
analysis.}

\item{treat_var}{This is a grouping variable. As a recommendation the treatment
groups should be used in the treat_var argument as this will provide a different color
for each of the treatments making it easier to identify.}

\item{Assay}{Name of the assay to be used for the pairwise analysis (default='normalized')}

\item{...}{Additional arguments to filter the analysis data by.}
}
\value{
Boxplots stratified by metabolites.
}
\description{
Creates boxplots for each metabolite within a specified subpathway.
}
\details{
.
}
