% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subpathway_lineplots.R
\name{subpathway_lineplots}
\alias{subpathway_lineplots}
\title{Subpathway Lineplots}
\usage{
subpathway_lineplots(MetPipe, subpathway, X, groupBy, ...)
}
\arguments{
\item{MetPipe}{MetPipe data object}

\item{subpathway}{Character value of the subpathway of interest. This is case
sensitive and must be in the chemical annotation file.}

\item{X}{This the the name of the variable in the meta data that is used for the
X axis of the line plots}

\item{groupBy}{This is a grouping variable. As a recommendation the treatment
groups should be used in the groupBy argument as this will provide a different color
for each of the treatments making it easier to identify.}

\item{...}{Additional arguments to filter the analysis data by.}
}
\value{
Line plots stratified by metabolite.
}
\description{
Create line plots for each metabolite within a subpathway.
}
