% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subpathway_lineplots.R
\name{subpathway_lineplots}
\alias{subpathway_lineplots}
\title{Subpathway Lineplots}
\usage{
subpathway_lineplots(
  data,
  subpathway,
  block_var,
  treat_var,
  Assay = "normalized",
  ...
)
}
\arguments{
\item{data}{SummarizedExperiment with Metabolon experiment data.}

\item{subpathway}{Character value of the subpathway of interest. This is case
sensitive and must be in the chemical annotation file.}

\item{block_var}{This the the name of the variable in the meta data that is used for the
X axis of the line plots. We recommend using the "block_var" variable from the
subpathway analyis.}

\item{treat_var}{This is a grouping variable. As a recommendation the treatment
groups should be used in the groupBy argument as this will provide a different color
for each of the treatments making it easier to identify.}

\item{Assay}{Name of the assay to be used for the pairwise analysis (default='normalized')}

\item{...}{Additional arguments to filter the analysis data by.}
}
\value{
Line plots stratified by metabolite.
}
\description{
Create line plots for each metabolite within a subpathway.
}
\examples{
#' # load data
dat = MetabolomicsPipeline::demoDat

################################################################################
### BoxPlots ###################################################################
################################################################################

subpathway_boxplots(dat, subpathway = "Lactoyl Amino Acid", block_var = TIME1,
                   treat_var = GROUP_NAME, Assay = "normalized",Gender =="Female")


################################################################################
## Line plots ##################################################################
################################################################################

# Set up data
dat$TIME1 <- as.numeric(factor(dat$TIME1,
                              levels = c("PreSymp","Onset","End")))
# Create line plots 
subpathway_lineplots(dat, subpathway = "Lactoyl Amino Acid",
                    block_var = TIME1,treat_var = GROUP_NAME, Assay = "normalized",Gender=="Female" )




}
